refeator : Ladder 클래스의 main메서드를 Main클래스를 만들어서 추출 했습니다. feat : Input 클래스와 Output클래스를 만들어서 입출력을 나누었습니다.
