+++ src/main/java/ladder/Output.java
@@ -1,0 +1,12 @@
+package ladder;
+
+public class Output {
+    public void printLadder(String[][] ladder) {
+        for (String[] strings : ladder) {
+            for (String string : strings) {
+                System.out.print(string);
+            }
+            System.out.println();
+        }
+    }
+}
